requirements.txt
flask
tensorflow
keras
opencv-python
numpy
pillow
matplotlib

